refactor: split style and logic
