+++ packages/website/src/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_IMAGE_URI = '/sample_image.jpg';
+export const CONTROLLER_BOUNDARY = 'controller-boundary';
