--- packages/website/src/consts.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_IMAGE_URI = '/sample_image.jpg';
